Fix delete user handling condition

+++ server/routes/users.js
@@ -78,7 +78,7 @@ export default (app) => {
       const user = await app.objection.models.user.query().findById(req.user.id);
       const notExecutedTasks = await user.$relatedQuery('executorTasks');
       const createdByUserTasks = await user.$relatedQuery('creatorTasks');
-      if (notExecutedTasks.length === 0 || createdByUserTasks.length === 0) {
+      if (notExecutedTasks.length === 0 && createdByUserTasks.length === 0) {
         req.logOut();
         await app.objection.models.user.query().deleteById(req.params.id);
         req.flash('info', i18next.t('flash.users.delete.success'));
